+++ clang/lib/Driver/ToolChains/Clang.cpp
@@ -5987,9 +5987,8 @@ void Clang::ConstructJob(Compilation &C, const JobAction &JA,
     }
   }
 
-  Args.AddLastArg(CmdArgs, options::OPT_fduplicate_mangled_name_EQ);
-  Args.addOptInFlag(CmdArgs, options::OPT_fmangle_capability_requirements,
-                    options::OPT_fno_mangle_capability_requirements);
+  Args.AddLastArg(CmdArgs,
+                  options::OPT_fthread_safety_capability_mangling_EQ);
 
   if (Arg *A = Args.getLastArg(options::OPT_fveclib)) {
     StringRef Name = A->getValue();